More fixes for MySQL peer (#1655)
Support rewrites for IN
