+++ nexus/Cargo.toml
@@ -4,6 +4,7 @@ members = [
   "catalog",
   "flow-rs",
   "parser",
+  "peer-ast",
   "peer-bigquery",
   "peer-connections",
   "peer-cursor",
@@ -19,9 +20,16 @@ members = [
 resolver = "2"
 
 [workspace.dependencies]
-chrono = { version = "0.4", default-features = false, features = ["serde", "std"] }
+chrono = { version = "0.4", default-features = false, features = [
+  "serde",
+  "std",
+] }
 dashmap = "5.0"
-rust_decimal = { version = "1", default-features = false, features = ["tokio-pg"] }
+rust_decimal = { version = "1", default-features = false, features = [
+  "tokio-pg",
+] }
 sqlparser = { git = "https://github.com/peerdb-io/sqlparser-rs.git", branch = "main" }
 tracing = "0.1"
-pgwire = { version = "0.22", default-features = false, features = ["server-api-ring"] }
+pgwire = { version = "0.22", default-features = false, features = [
+  "server-api-ring",
+] }
