+++ nexus/peer-ast/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "peer-ast"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = "1.0"
+pt = { path = "../pt" }
+rust_decimal.workspace = true
+sqlparser.workspace = true
