+++ nexus/peer-bigquery/Cargo.toml
@@ -10,6 +10,7 @@ anyhow = "1.0"
 async-trait = "0.1"
 chrono.workspace = true
 futures = { version = "0.3.28", features = ["executor"] }
+peer-ast = { path = "../peer-ast" }
 peer-cursor = { path = "../peer-cursor" }
 peer-connections = { path = "../peer-connections" }
 pgwire.workspace = true
